1. Added support for checking `@since` tags for class, trait, interface, and enum declarations. 2. Added support for checking `@since` tags for class and trait properties; 3. Added support for checking `@since` tags for class, interface, and trait method declarations; 4. Added support for checking `@since` tags for WordPress functions that invoke hooks: `do_action()`, `do_action_ref_array()`, `apply_filters()`, `apply_filters_ref_array()`. 5. Refactored unit tests; duplicate code moved to the new `GutenbergCS\Gutenberg\Tests\AbstractSniffUnitTest` class.
