+++ phpcs.xml.dist
@@ -172,8 +172,11 @@
 			</property>
 		</properties>
 	</rule>
-	<rule ref="Gutenberg.Commenting.FunctionCommentSinceTag">
+	<rule ref="Gutenberg.Commenting.SinceTag">
 		<!-- The sniff ensures that functions have a valid @since tag but skips checking experimental blocks. -->
 		<include-pattern>/packages/block-library/src/.+/*\.php$</include-pattern>
+		<include-pattern>/packages/block-serialization-default-parser/.+/*\.php$</include-pattern>
+		<include-pattern>/packages/widgets/src/blocks/legacy-widget/index\.php$</include-pattern>
+		<include-pattern>/packages/widgets/src/blocks/widget-group/index\.php$</include-pattern>
 	</rule>
 </ruleset>
